Extract dataset creation logic into shared service
- Create shared dataset-service.ts that handles file upload, validation, and variable metadata extraction
- Update server action to use shared service while maintaining auth checks
- Update seed script to use shared service for consistent dataset creation with variables
- Add support for predefined IDs in service to enable deterministic seeding
- Remove code duplication between server action and seeding functionality
